# backend/swagger.yaml
openapi: 3.0.1
info:
  title: FluxPilot API
  description: "REST API платформы FluxPilot для управления workflow, их версиями\
    , снапшотами, триггерами, запусками и A/B-аналитикой."
  version: 1.0.0
servers:
- url: /v1
  description: Основной API сервер
...
paths:
  /workflows/{workflowId}:
    get:
      tags:
      - Workflows
      operationId: get
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Workflow"
    put:
      tags:
      - Workflows
      operationId: updateMeta
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowMetaUpdate"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowMeta"
    delete:
      tags:
      - Workflows
      operationId: delete
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
  /workflow-versions/{versionId}/graph:
    put:
      tags:
      - Versions
      operationId: putGraph
      parameters:
      - name: versionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowGraph"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowGraph"
  /workflows:
    get:
      tags:
      - Workflows
      operationId: list
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowMeta"
    post:
      tags:
      - Workflows
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowCreateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{workflowId}/versions:
    get:
      tags:
      - Workflows
      summary: Список версий workflow
      operationId: listVersions
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowVersion"
    post:
      tags:
      - Workflows
      summary: Создать именованную версию
      description: Фиксирует текущую ревизию как именованную версию (тег опционален)
      operationId: createVersion
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkflowVersionCreateRequest"
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
  /workflows/{workflowId}/versions/{versionId}/restore:
    post:
      tags:
      - Workflows
      summary: Откат к версии
      description: Создаёт новую (append-only) ревизию с графом указанной версии
      operationId: restoreVersion
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: versionId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowGraph"
  /workflows/{workflowId}/snapshots:
    get:
      tags:
      - Snapshots
      operationId: list_1
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkflowSnapshot"
    post:
      tags:
      - Snapshots
      operationId: create_1
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSnapshotRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowSnapshot"
  /workflows/{workflowId}/snapshots/{snapshotId}/restore:
    post:
      tags:
      - Snapshots
      operationId: restore
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WorkflowGraph"
...
  /workflows/{workflowId}/snapshots/{snapshotId}:
    delete:
      tags:
      - Snapshots
      operationId: delete_1
      parameters:
      - name: workflowId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: snapshotId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    WorkflowMetaUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    WorkflowMeta:
      required:
      - createdAt
      - id
      - isDemo
      - name
      - nodesCount
      - updatedAt
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        isDemo:
          type: boolean
        nodesCount:
          type: integer
          format: int32
        createdAt:
          type: string
        updatedAt:
          type: string
    Connection:
      required:
      - id
      - source
      - target
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        sourceHandle:
          type: string
        targetHandle:
          type: string
        variant:
          type: string
    JsonNode:
      type: object
    Node:
      required:
      - id
      - type
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        position:
          $ref: "#/components/schemas/Position"
        data:
          $ref: "#/components/schemas/NodeData"
    NodeData:
      type: object
      properties:
        label:
          type: string
        config:
          $ref: "#/components/schemas/JsonNode"
        abConfig:
          $ref: "#/components/schemas/JsonNode"
    Position:
      required:
      - x
      - "y"
      type: object
      properties:
        x:
          type: number
          format: double
        "y":
          type: number
          format: double
    WorkflowGraph:
      required:
      - connections
      - nodes
      - versionId
      type: object
      properties:
        versionId:
          type: string
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/Node"
        connections:
          type: array
          items:
            $ref: "#/components/schemas/Connection"
    WorkflowCreateRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    Workflow:
      required:
      - graph
      - meta
      type: object
      properties:
        meta:
          $ref: "#/components/schemas/WorkflowMeta"
        graph:
          $ref: "#/components/schemas/WorkflowGraph"
    WorkflowVersionCreateRequest:
      type: object
      properties:
        versionTag:
          type: string
    WorkflowVersion:
      required:
      - createdAt
      - id
      - workflowId
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        tag:
          type: string
        createdAt:
          type: string
    CreateSnapshotRequest:
      required:
      - name
      type: object
      properties:
        name:
          type: string
        description:
          type: string
    WorkflowSnapshot:
      required:
      - createdAt
      - id
      - name
      - workflowId
      type: object
      properties:
        id:
          type: string
        workflowId:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
...